Do not swallow errors in simple query.
This fixes a problem where we might have warnings coming before errors
in multiple resultset queries, when executing with `simple_query`.

Example:

```
CREATE UNIQUE INDEX [X] ON [table] ([column]);
```

In this case, if our column breaks unique validations, we'd expect
`simple_query` to return an error.

If the column is of type `varchar` and the length is above threshold,
the first token we get back is a warning of index size. We used to stop
fetching data at this point, and our query would be successful even
though we actually had an error in the wire.