Try out all resolved IP addresses (#124)
On SQL Browser, we previously just tried the first resolved IP, and
failed with connection refused if it didn't work. Now there's a docker
bug that doesn't allow us to connect with `::1`, meaning using
`localhost` with the docker image would tehn try to connect to `::1`,
fail and stop the world.

Now we look into other addresses too, and try again if the first ones
didn't work.