Refactor stream handling to something more rusty (#166)
* Refactor stream handling to something more rusty

Instead of needing to call `next_resultset()`, we now stream `QueryItem`
enums, that can be either column metadata or a row. Additionally, the
metadata and row strucs hold a `result_index` to distict the items from
different result sets.

A stream always starts with a metadata item with the column data,
followed by rows from that result set. If having more than one result in
the stream, a new metadata item will appear from the stream, followed by
a different set of rows.

* Add `columns` to `QueryResult`

* Add `into_row_stream` into `QueryResult`

* Get rid of QueryResult

* Fix broken example

* Docs wording

* Format Rust code using rustfmt

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>