Dynamic query interface
Introduces a new API to perform queries that bind dynamic parameters, making
them easier without needing to cast dyn traits out of the data before using.

The public API is the `Query` struct:

- Initialize with the query string
- Bind parameters one by one
- Either query or execute with a connection

Extends the `IntoSql` implementations to all types, making the binding
by-value, but allowing using references if needed.

In the end, might be we want to remove the `ToSql` completely, but not yet.