Additional option to add custom-ca to root certificates
fixes prisma/tiberius#197

Next to trust_cert another option is now available to instead add another root-certificate in pem or der format.
The new option is also supported as ado-/jdbc-connection-strings.

Certificate generation scripts have been added to generate a self-signing ca and a server certificate.
The server certificate and ca is added to the two MSSQL instances in Docker-Compose.

Enabled Buildkit to make use of chmod in Dockerfile

Do not force the encryption on the server-side so that plain-text tests still succeed