Propagate errors correctly
We have to consume the whole token stream to the end even if having
an error. Error is stored to the stream, and further tokens are all
handled before responding to the user.

This way a possible rollback can zero the transaction descriptor, and
the connection can be used for further queries without an error.